# Conflicts:
#	social-network/src/main/java/com/skillbox/socialnetwork/service/FriendshipService.java
#	social-network/src/main/java/com/skillbox/socialnetwork/service/PersonService.java
#	social-network/src/main/java/com/skillbox/socialnetwork/service/RedisService.java

+++ social-network/src/main/java/com/skillbox/socialnetwork/service/TagService.java
@@ -61,15 +61,16 @@ public class TagService {
         Tag tag = tagRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Tag is not exist"));
         Set<Post> postsWithTag = postRepository.findPostsByTag(tag.getTag());
-        postsWithTag.forEach(p -> {
-            p.getTags().remove(tag);
-            postRepository.save(p);
-        });
-        tagRepository.deleteById(id);
         AccountResponse response = new AccountResponse();
-        response.setTimestamp(Instant.now());
-        response.setError("nothing");
+        if (postsWithTag.isEmpty()) {
+            tagRepository.deleteById(id);
+            response.setError("nothing");
+        } else {
+            response.setError("tag use in another posts");
+        }
         response.setData(Map.of("message", "ok"));
+        response.setTimestamp(Instant.now());
+
         return response;
     }
 }
